requirements.txt
click==8.1.7
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==0.16.1
grpclib>=0.4.0,<0.5.0
grpcio-tools==1.59.0
protobuf